refactor: convert reportUser.js to ts
